Flush CSVs to Redshift after hitting a preset threshold
To avoid having large CSV files to be copied over to Redshift

+++ airbyte-integrations/connectors/destination-redshift-py/destination_redshift_py/destination.py
@@ -21,7 +21,7 @@ from destination_redshift_py.csv_writer import CSVWriter
 from destination_redshift_py.jsonschema_to_tables import JsonToTables, PARENT_CHILD_SPLITTER
 from destination_redshift_py.s3_objects_manager import S3ObjectsManager
 from destination_redshift_py.stream import Stream
-from destination_redshift_py.table import AIRBYTE_EMITTED_AT, AIRBYTE_ID_NAME
+from destination_redshift_py.table import AIRBYTE_EMITTED_AT, AIRBYTE_ID_NAME, Table
 
 logger = logging.getLogger("airbyte")
 
@@ -92,7 +92,7 @@ class DestinationRedshiftPy(Destination):
         """
         for message in input_messages:
             if message.type == Type.STATE:
-                self._flush()
+                self._flush(force=True)
                 yield message
             elif message.type == Type.RECORD:
                 nested_record = DotMap({message.record.stream: message.record.data})
@@ -112,7 +112,8 @@ class DestinationRedshiftPy(Destination):
                 emitted_at = message.record.emitted_at
 
                 for node in nodes:
-                    table = stream.final_tables[PARENT_CHILD_SPLITTER.join(node)]
+                    final_table = stream.final_tables[PARENT_CHILD_SPLITTER.join(node)]
+                    staging_table = stream.staging_tables[PARENT_CHILD_SPLITTER.join(node)]
 
                     def get_records(parents: Union[DotMap, List[DotMap]], method: str) -> List[DotMap]:
                         if not isinstance(parents, list):
@@ -125,8 +126,8 @@ class DestinationRedshiftPy(Destination):
                                 children = [children]
 
                             for child_item in children:
-                                if child_item and table.references and method == node[-1]:
-                                    child_item[table.reference_key.name] = parent_item[AIRBYTE_ID_NAME]
+                                if child_item and final_table.references and method == node[-1]:
+                                    child_item[final_table.reference_key.name] = parent_item[AIRBYTE_ID_NAME]
 
                                 children_records.append(child_item)
 
@@ -138,18 +139,26 @@ class DestinationRedshiftPy(Destination):
                     if records:
                         # Choose primary keys (other than the Airbyte auto generated ID). If there are no primary keys (except the auto
                         # generated Airbyte ID, this happens with children), then choose all the fields as hash keys to generate the ID.
-                        hashing_keys = [pk for pk in table.primary_keys if pk != AIRBYTE_ID_NAME] or table.field_names
+                        hashing_keys = [pk for pk in final_table.primary_keys if pk != AIRBYTE_ID_NAME] or final_table.field_names
 
                         self._assign_id_and_emitted_at(records=records, hashing_keys=hashing_keys, emitted_at=emitted_at)
 
                         # Assign only the table fields and drop other fields
                         records = [
-                            DotMap([field_name, record[field_name] or None] for field_name in table.field_names) for record in records
+                            DotMap([field_name, record[field_name] or None] for field_name in final_table.field_names) for record in records
                         ]
 
-                        self.csv_writers[table.name].write(records)
+                        csv_writer = self.csv_writers[final_table.name]
+                        csv_writer.write(records)
 
-        self._flush()
+                        self._flush_csv_writer_to_destination(
+                            csv_writer=csv_writer,
+                            final_table=final_table,
+                            staging_table=staging_table,
+                            mode=stream.destination_sync_mode
+                        )
+
+        self._flush(force=True)
 
     @staticmethod
     def _assign_id_and_emitted_at(records: List[DotMap], emitted_at: int, hashing_keys: List[str]):
@@ -221,7 +230,7 @@ class DestinationRedshiftPy(Destination):
             password=config.get("password")
         )
 
-    def _get_connection(self, autocommit=False):
+    def _get_connection(self, autocommit: bool = False):
         connection = self.connection_pool.getconn()
         connection.autocommit = autocommit
 
@@ -251,32 +260,42 @@ class DestinationRedshiftPy(Destination):
     def _initialize_iam_role_arn(self, iam_role_arn: str):
         self.iam_role_arn = iam_role_arn
 
-    def _flush(self):
-        connection = self._get_connection()
-        cursor = connection.cursor()
-
+    def _flush(self, force: bool = False):
         for stream in self.streams.values():
             for key, final_table in stream.final_tables.items():
-                csv_writer = self.csv_writers[final_table.name]
-                temporary_gzip_file = csv_writer.flush_gzipped()
+                self._flush_csv_writer_to_destination(
+                    csv_writer=self.csv_writers[final_table.name],
+                    final_table=final_table,
+                    staging_table=stream.staging_tables[key],
+                    mode=stream.destination_sync_mode,
+                    force=force
+                )
 
-                if temporary_gzip_file:
-                    s3_full_path = self.s3_object_manager.upload_file_to_s3(temporary_gzip_file.name)
-                    CSVWriter.delete_gzip_file(temporary_gzip_file)
+    def _flush_csv_writer_to_destination(self, csv_writer: CSVWriter, final_table: Table, staging_table: Table, mode: DestinationSyncMode,
+                                         force: bool = False):
+        if csv_writer.is_flushable() or force:
+            rows_count = csv_writer.rows_count()
+            temporary_gzip_file = csv_writer.flush_gzipped()
 
-                    if stream.destination_sync_mode in [DestinationSyncMode.append, DestinationSyncMode.overwrite]:
-                        copy_statement = final_table.copy_csv_gzip_statement(iam_role_arn=self.iam_role_arn, s3_full_path=s3_full_path)
-                        cursor.execute(copy_statement)
-                    else:
-                        staging_table = stream.staging_tables[key]
+            if temporary_gzip_file:
+                logger.info(f"Flushing {rows_count} to destination")
 
-                        copy_statement = staging_table.copy_csv_gzip_statement(iam_role_arn=self.iam_role_arn, s3_full_path=s3_full_path)
-                        cursor.execute(copy_statement)
+                connection = self._get_connection()
+                cursor = connection.cursor()
 
-                        upsert_statements = final_table.upsert_statements(staging_table=staging_table)
-                        cursor.execute(upsert_statements)
+                s3_full_path = self.s3_object_manager.upload_file_to_s3(temporary_gzip_file.name)
+                CSVWriter.delete_gzip_file(temporary_gzip_file)
 
-                    self.s3_object_manager.delete_file_from_s3(s3_full_path)
+                if mode in [DestinationSyncMode.append, DestinationSyncMode.overwrite]:
+                    copy_statement = final_table.copy_csv_gzip_statement(iam_role_arn=self.iam_role_arn, s3_full_path=s3_full_path)
+                    cursor.execute(copy_statement)
+                else:
+                    copy_statement = staging_table.copy_csv_gzip_statement(iam_role_arn=self.iam_role_arn, s3_full_path=s3_full_path)
+                    cursor.execute(copy_statement)
+
+                    upsert_statements = final_table.upsert_statements(staging_table=staging_table)
+                    cursor.execute(upsert_statements)
+
+                self.s3_object_manager.delete_file_from_s3(s3_full_path)
 
                 connection.commit()
-
